Update slug format for post type
Posts no longer need to specify a full url for the resulting html.
Rather, the prefix portion of the url will be inferred from the
post's location in the source filesystem. The post slug now simply
represents a url friendly version of the title.

Additionally, updates all sample posts to use the new slug format.